extract analytics related hooks into one hook
